Se hace multi-tenant

+++ ShingekiNoAPP/BarberShopAPI/Negocio/BusinessEntities/CashMovement.cs
@@ -8,7 +8,7 @@ namespace Business.BusinessEntities
 {
     public class CashMovement : BaseEntity
     {
-        public int CashSessionId { get; set; }
+        public long CashSessionId { get; set; }
         public string Type { get; set; } // "IN" (Ingreso) o "OUT" (Egreso/Gasto)
         public decimal Amount { get; set; }
         public string Description { get; set; }
